+++ backend/src/main/java/com/morak/back/appointment/domain/appointment/timeconditions/appointmentdateperiod/AppointmentDate.java
@@ -1,4 +1,4 @@
-package com.morak.back.appointment.domain.timeconditions.appointmentdateperiod;
+package com.morak.back.appointment.domain.appointment.timeconditions.appointmentdateperiod;
 
 import com.morak.back.appointment.exception.AppointmentDomainLogicException;
 import com.morak.back.core.exception.CustomErrorCode;
