+++ backend/src/main/java/com/morak/back/appointment/domain/appointment/timeconditions/appointmenttimeperiod/AppointmentTime.java
@@ -1,4 +1,4 @@
-package com.morak.back.appointment.domain.timeconditions.appointmenttimeperiod;
+package com.morak.back.appointment.domain.appointment.timeconditions.appointmenttimeperiod;
 
 import com.morak.back.appointment.exception.AppointmentDomainLogicException;
 import com.morak.back.core.exception.CustomErrorCode;
