+++ backend/src/main/java/com/morak/back/appointment/ui/dto/AppointmentCreateRequest.java
@@ -1,7 +1,7 @@
 package com.morak.back.appointment.ui.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.morak.back.appointment.domain.Appointment;
+import com.morak.back.appointment.domain.appointment.Appointment;
 import com.morak.back.auth.domain.Member;
 import com.morak.back.core.domain.Code;
 import com.morak.back.core.domain.times.Times;
