+++ backend/src/main/java/com/morak/back/core/domain/slack/FormattableData.java
@@ -1,6 +1,6 @@
 package com.morak.back.core.domain.slack;
 
-import com.morak.back.appointment.domain.Appointment;
+import com.morak.back.appointment.domain.appointment.Appointment;
 import com.morak.back.core.support.Generated;
 import com.morak.back.poll.domain.Poll;
 import java.time.LocalDateTime;
