+++ backend/src/main/java/com/morak/back/poll/domain/PollItem.java
@@ -25,21 +25,18 @@ import lombok.NoArgsConstructor;
 @Getter
 public class PollItem extends BaseEntity {
 
+    @OneToMany(mappedBy = "pollItem", cascade = CascadeType.ALL, orphanRemoval = true)
+    private final List<PollResult> pollResults = new ArrayList<>();
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn
     @NotNull(message = "poll 은 null 일 수 없습니다.")
     private Poll poll;
-
     @NotBlank(message = "subject 는 blank 일 수 없습니다.")
     @Size(min = 1, max = 255, message = "투표 항목 주제의 길이는 1~255자여아 합니다.")
     private String subject;
-
-    @OneToMany(mappedBy = "pollItem", cascade = CascadeType.ALL, orphanRemoval = true)
-    private final List<PollResult> pollResults = new ArrayList<>();
 
     @Builder
     private PollItem(Long id, Poll poll, String subject) {
