+++ backend/src/test/java/com/morak/back/appointment/domain/AppointmentRepositoryTest.java
@@ -3,7 +3,9 @@ package com.morak.back.appointment.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.morak.back.appointment.domain.Appointment.AppointmentBuilder;
+import com.morak.back.appointment.domain.appointment.Appointment;
+import com.morak.back.appointment.domain.appointment.Appointment.AppointmentBuilder;
+import com.morak.back.appointment.domain.appointment.AppointmentRepository;
 import com.morak.back.appointment.domain.availabletime.AvailableTime;
 import com.morak.back.appointment.domain.availabletime.AvailableTimeRepository;
 import com.morak.back.auth.domain.Member;
