+++ backend/src/test/java/com/morak/back/appointment/domain/AvailableTimeRepositoryTest.java
@@ -3,6 +3,8 @@ package com.morak.back.appointment.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import com.morak.back.appointment.domain.appointment.Appointment;
+import com.morak.back.appointment.domain.appointment.AppointmentRepository;
 import com.morak.back.appointment.domain.availabletime.AvailableTime;
 import com.morak.back.appointment.domain.availabletime.AvailableTimeRepository;
 import com.morak.back.auth.domain.Member;
